pyproject.toml
[project]
name = "stark"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "minio>=7.2.16",
    "openpyxl>=3.1.5",
    "oracledb>=3.3.0",
    "pandas>=2.3.1",
    "plotly>=6.3.0",
    "pre-commit>=4.3.0",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
    "streamlit>=1.48.1",
]
